New updated migration
