+++ db/moonstreamdb/models.py
@@ -6,6 +6,7 @@ from sqlalchemy import (
     Integer,
     ForeignKey,
     MetaData,
+    Numeric,
     Text,
     VARCHAR,
 )
@@ -49,16 +50,13 @@ class EthereumBlock(Base):  # type: ignore
     __tablename__ = "ethereum_blocks"
 
     block_number = Column(
-        BigInteger,
-        primary_key=True,
-        unique=True,
-        nullable=False,
+        BigInteger, primary_key=True, unique=True, nullable=False, index=True
     )
     difficulty = Column(BigInteger)
     extra_data = Column(VARCHAR(128))
     gas_limit = Column(BigInteger)
     gas_used = Column(BigInteger)
-    hash = Column(VARCHAR(256))
+    hash = Column(VARCHAR(256), index=True)
     logs_bloom = Column(VARCHAR(1024))
     miner = Column(VARCHAR(256))
     nonce = Column(VARCHAR(256))
@@ -79,24 +77,22 @@ class EthereumTransaction(Base):  # type: ignore
     __tablename__ = "ethereum_transactions"
 
     hash = Column(
-        VARCHAR(256),
-        primary_key=True,
-        unique=True,
-        nullable=False,
+        VARCHAR(256), primary_key=True, unique=True, nullable=False, index=True
     )
     block_number = Column(
         BigInteger,
         ForeignKey("ethereum_blocks.block_number", ondelete="CASCADE"),
         nullable=False,
+        index=True,
     )
     from_address = Column(VARCHAR(256), index=True)
     to_address = Column(VARCHAR(256), index=True)
-    gas = Column(Text)
-    gas_price = Column(Text)
+    gas = Column(Numeric(precision=78, scale=0), index=True)
+    gas_price = Column(Numeric(precision=78, scale=0), index=True)
     input = Column(Text)
     nonce = Column(VARCHAR(256))
     transaction_index = Column(BigInteger)
-    value = Column(Text)
+    value = Column(Numeric(precision=78, scale=0), index=True)
 
     indexed_at = Column(
         DateTime(timezone=True), server_default=utcnow(), nullable=False
@@ -107,24 +103,22 @@ class EthereumPendingTransaction(Base):  # type: ignore
     __tablename__ = "ethereum_pending_transactions"
 
     hash = Column(
-        VARCHAR(256),
-        primary_key=True,
-        unique=True,
-        nullable=False,
+        VARCHAR(256), primary_key=True, unique=True, nullable=False, index=True
     )
     block_number = Column(
         BigInteger,
         ForeignKey("ethereum_blocks.block_number", ondelete="CASCADE"),
         nullable=False,
+        index=True,
     )
     from_address = Column(VARCHAR(256), index=True)
     to_address = Column(VARCHAR(256), index=True)
-    gas = Column(Text)
-    gas_price = Column(Text)
+    gas = Column(Numeric(precision=78, scale=0), index=True)
+    gas_price = Column(Numeric(precision=78, scale=0), index=True)
     input = Column(Text)
     nonce = Column(VARCHAR(256))
     transaction_index = Column(BigInteger)
-    value = Column(Text)
+    value = Column(Numeric(precision=78, scale=0), index=True)
 
     indexed_at = Column(
         DateTime(timezone=True), server_default=utcnow(), nullable=False
@@ -138,7 +132,7 @@ class ESDFunctionSignature(Base):
 
     __tablename__ = "esd_function_signatures"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(Integer, primary_key=True, unique=True, nullable=False, index=True)
     text_signature = Column(Text, nullable=False)
     hex_signature = Column(VARCHAR(10), nullable=False)
     created_at = Column(
@@ -153,7 +147,7 @@ class ESDEventSignature(Base):
 
     __tablename__ = "esd_event_signatures"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(Integer, primary_key=True, unique=True, nullable=False, index=True)
     text_signature = Column(Text, nullable=False)
     hex_signature = Column(VARCHAR(66), nullable=False)
     created_at = Column(
